Add SQL comment to original lock string for PostgreSQL
The PostgreSQL implementation requires translating the user-specified
lock string into a pair of 32 bit numbers, which makes the lock and
unlock statements opaque in the SQL logs.

To aid debugging, add a SQL comment containing the original lock string.
The comment will be removed from the input by the PostgreSQL parser, but
it will still show up in logs.